Optimize table to trigger liquid clustering before test [databricks] (#13371)
Fixes #13370.

### Description

For Delta IO, enabling liquid clustering for a table does not
automatically trigger clustering as data is appended. [The table should
be optimized to do
that](https://docs.delta.io/delta-clustering/#how-to-trigger-clustering).
Currently, our tests for liquid clustering in
`delta_lake_clustered_reads_test.py` create a table with liquid
clustering enabled, but do not trigger the clustering. As for databricks
delta, we have the same issue even though [Databricks delta can
automatically trigger the clustering if it satisfies certain
condition](https://docs.databricks.com/gcp/en/delta/clustering#write-data-to-a-clustered-table),
because our test data doesn't seem to satisfy that condition. This PR
adds an optimize command after data is appended to the test table.

The tests in `delta_lake_liquid_clustereing_test.py` are fine as they
test only fallbacks.

Performance test result will be posted in #13370.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [x] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

Signed-off-by: Jihoon Son <ghoonson@gmail.com>