Add support for iceberg bucket expression of int and long types. (#13397)
Related to #13382.

### Description

This pr implements gpu version of iceberg bucket expression for int and
long types, utilizing existing jni implementation.

Other types such as binary, string and decimal are not implemented yet,
we need add implemementations of gpu kernel for them.

Added ut.

### Checklists


- [x] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: liurenjie1024 <liurenjie2008@gmail.com>