Fix docstring typo in collect_data_or_exception [skip ci] (#13246)
This is a very small few-character PR, to fix text in a doctring that
github copilot caught.

### Description

I am trying to address this comment by copilot:
https://github.com/NVIDIA/spark-rapids/pull/13236/files/36c6d87b9507f6328bc6f36b855f5c622df59088#r2251688875

### Checklists

This PR has:

- [x] added documentation for new or modified features or behaviors.
- [ ] updated the license in the source code files when it is required.
- [ ] added new tests or modified existing tests to cover new code
paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)

Please select one of the following options:
- [ ] Performance testing has been performed and its results are added
in the PR description.
- [ ] An issue is filed for performance testing and its link is added in
the PR description. (Select this if performance testing will not be
completed before the PR is submitted.)

---------

Signed-off-by: Alessandro Bellina <abellina@nvidia.com>