Allows AQE to fall back to the CPU for non-reused broadcast (#13367)
Fixes #13360

### Description

The original code assumed that AQE would only ever fall back to the CPU
for a reused exchange. That was wrong and the code now does the right
thing in those cases.

### Checklists

- [X] This PR has added documentation for new or modified features or
behaviors.
   - No need it is fixing a bug it is not a feature
- [X] This PR has added new tests or modified existing tests to cover
new code paths.
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.
- The performance is the same as it would be for any other broadcast fix
up. It is not perfect and only deals with a broadcast that should
typically be a few MiB.

---------

Signed-off-by: Robert (Bobby) Evans <bobby@apache.org>