Add rule to override StaticInvoke for iceberg (#13438)
Closes #13390 .

### Description

When writing to iceberg partitioned table, iceberg will add some
`StaticInvoke` expression to compute partition value for
sorting/partitioning purpose. We need to override this expression so
that it could be accelated by gpu.

### Checklists


- [x] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: liurenjie1024 <liurenjie2008@gmail.com>