Support InMemoryTableScan with AQE for Spark 3.5.2+ (#13434)
Fixes https://github.com/NVIDIA/spark-rapids/issues/8678

### Description
In this PR, we added support to accelerate operator InMemoryTableScan on
GPU for Spark versions 3.5.2 and above.

InMemoryTableScanExec was wrapped in TableCacheQueryStageExec in
Spark-3.5.0 https://issues.apache.org/jira/browse/SPARK-42101 . With
this change, InMemoryTableScanExec couldn't run on GPU as before. A
trait for InMemoryTableScanExec was added in this PR
https://github.com/apache/spark/pull/45525 so that it could be extended
by the plugin. The PR was merged in branch-3.5 and part of 3.5.2
release.

This PR:
Keeps InMemoryTableScan disabled-by-default on Spark 3.5.0 and 3.5.1 due
to missing InMemoryTableScanLike trait.
Adds shims/utilities for version-specific behavior and comprehensive
integration tests with version-aware expectations.

#### Behavior by Spark version:
- 3.2.0 - 3.4.x: InMemoryTableScan can run on GPU (existing behavior).
- 3.5.0-3.5.1: InMemoryTableScan disabled by default. So this operator
falls back to CPU.
- 3.5.2+: InMemoryTableScan runs on GPU and is AQE-compatible via
`InMemoryTableScanLike`.

#### Testing:
- Added integration tests updatedto reflect version-specific
expectations.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Niranjan Artal <nartal@nvidia.com>