Add Partial support for Deletion Vector for PERFILE Reader Type (#13439)
Fixes #12575

### Description

- Today, if a user tries to read a table with deletion vectors, the
expression falls back to the CPU
- This PR adds support for reading the deletion vectors on the GPU. The
following conf need to be set in order for this to work:
- `spark.databricks.delta.deletionVectors.useMetadataRowIndex` needs to
be set to false otherwise we will fallback due to hidden metadata fields
that we don't support on the GPU.
- `spark.rapids.sql.format.parquet.reader.type=PERFILE` needs to be set
as the current change only affects the perfile reader
- The fix involves updating the read function that takes in a
`PartitionedFile` and returns a ColumnarBatch. The change involves
getting the associated Deltion Vector and creating the necessary columns
to filter out the deleted rows.
- An integration test was added that exercises the path for reading the
table with deletion vectors.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [x] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Raza Jafri <raza.jafri@gmail.com>