Add support for Iceberg insert integration tests. (#13460)
Fixes #13380 .

### Description

This pr adds integration tests for iceberg insert statement.

### Checklists

- [x] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: liurenjie1024 <liurenjie2008@gmail.com>