Refine column size estimation based on previous batch for Host2Gpu (#13499)
Fixes #13449.

### Description

When we move data from host to GPU, we can concat multiple input batches
to reduce the number of data moves. For concatting batches, we use a
row-based limit which is based on schema size estimation. The row limit
is computed based on the user-configurable bytes limit and the schema
size estimate. As for the schema size estimation, for the first output
batch, we use a sort of naive approach as we don't have much information
about actual data size. This is implemented in
[here](https://github.com/NVIDIA/spark-rapids/blob/branch-25.10/sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuBatchUtils.scala#L97-L129).
This naive approach can be problematic especially for variable-sized
types like the string or the blob as it uses the default size of those
types which is some fixed number. As such, the inaccurate size
estimation may let too many rows be in the first output batch, which
causes some overflow error while allocating GPU memory. Note that we
don't have the same issue for subsequent batches other than the first
one as we refine the limit based on the actual output batch size.

The fix proposed in this PR is to refine the limit as we go based on the
input batch size. As a new batch is added to concat, we measure the
actual byte size to concat. The measuring is done as part of the
existing logic that iterates all rows in all columns. That being said,
we still have the same inaccurate size estimation issue for the very
first input batch. If the first input batch is very large, we will still
see the overflow error.

```sql
create table lineitem_test_delete using delta TBLPROPERTIES ('delta.enableDeletionVectors' = 'true') as select * from lineitem;

delete from lineitem_test_delete where l_discount = 0.01;
```

Since the overflow error occurs only with a reasonably large data, I
tested this change manually using the queries used in #13449 (also shown
above). I used the TPC-H lineitem table at sf=100. The delete ran
successfully.

```sql
create table lineitem_test_delete using delta TBLPROPERTIES ('delta.enableDeletionVectors' = 'true') as select * from lineitem limit 16519100;

delete from lineitem_test_delete where l_discount = 0.01;
```

I also ran some performance tests to measure the overhead of this
change. I used the above query and the same table. I did not observe any
performance change between the 25.10 branch
(837ba27df6b92120a64d9dad6c33679bc422019f) and this PR.

Configs used:

```
export SPARK_CONF=("--master" "local[16]"
                   "--conf" "spark.driver.maxResultSize=2GB"
                   "--conf" "spark.driver.memory=50G"
                   "--conf" "spark.executor.memory=16G"
                   "--conf" "spark.sql.files.maxPartitionBytes=2gb"
                   "--conf" "spark.sql.adaptive.enabled=true"
                   "--conf" "spark.plugins=com.nvidia.spark.SQLPlugin"
                   "--conf" "spark.rapids.memory.host.spillStorageSize=16G"
                   "--conf" "spark.rapids.memory.pinnedPool.size=8g"
                   "--conf" "spark.rapids.sql.concurrentGpuTasks=3"
                   "--packages" "io.delta:delta-spark_2.12:3.3.1"
                   "--conf" "spark.sql.extensions=io.delta.sql.DeltaSparkSessionExtension"
                   "--conf" "spark.sql.catalog.spark_catalog=org.apache.spark.sql.delta.catalog.DeltaCatalog"
                   "--conf" "spark.driver.extraClassPath=$SPARK_RAPIDS_PLUGIN_JAR:$NDS_LISTENER_JAR"
                   "--conf" "spark.executor.extraClassPath=$SPARK_RAPIDS_PLUGIN_JAR:$NDS_LISTENER_JAR")
```

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [x] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>