Switch rounding calls to using new API (#13497)
This changes the plugin to call the new spark-rapids-jni API for
rounding rather than the old cudf API. The rounding of floating point
functionality will be removed from cudf so we've moved the float
rounding (and the API) into spark-rapids-jni.

This is part of resolving [sr-jni issue
3585](https://github.com/NVIDIA/spark-rapids-jni/issues/3585)

[This PR](https://github.com/rapidsai/cudf/pull/20110) removes the old
java code from cudf, and will be merged in cudf 25.12.
[This PR](https://github.com/NVIDIA/spark-rapids-jni/pull/3770) moves
the floating rounding into spark-rapids-jni and changes the API. This
will be merged in 25.10 and is a dependency of this PR.

Note that I have tested this code applying both of the above PRs and all
of the spark-rapids integration tests pass.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Paul Mattione <pmattione@nvidia.com>