Avoid throwing ClassNotFoundExceptions when checking isSupportedFormat with Delta Lake versions < 2.0.1 (#13558)
Fixes #13519 

### Description

We recently encountered an issue when running the plugin with Delta Lake
< 1.1.0, which resulted in a crash. Although we don't officially support
Delta Lake versions < 2.0.1, this regression is unacceptable; this PR
addresses it and reverts to the old behavior.

The PR was tested by running the NDS benchmark against Delta Lake 1.1.0
and 1.2.0 with Spark 3.2.0

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Raza Jafri <raza.jafri@gmail.com>