Add iceberg tests of insert into with values (#13545)
### Description

When insert into iceberg table, if the input plan can't be converted to
`GpuExec`, we should convert wrap it with `GpuRowToColumnarExec`

### Checklists


- [x] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

Signed-off-by: Ray Liu <liurenjie2008@gmail.com>