Fix read option name for change data feed for Delta [databricks] (#13532)
Fixes #12796

### Description

We are currently using "readChangeDataFeed" as a read option name to
read a table with change data feed in our tests. However, this option
name is invalid. We should use "readChangeFeed" instead. This PR fixes
the option name and adds asserts to validate the schema of the data
read.

I manually ran `delta_lake_delete_test.py`, `delta_lake_update_test.py`,
and `delta_lake_merge_test.py`, which are impacted by this change.
`delta_lake_low_shuffle_merge_test.py` is also impacted as well, but it
does not run with Delta 3.3. I used Spark 3.5.5 and Delta 3.3.0 for my
testing.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>