Upgrade UCX to 1.19-rc1 (#13572)
Fixes https://github.com/NVIDIA/spark-rapids/issues/13266

### Description
Update all ucx version references in the repo to point to 1.19.1-rc1,
and update the corresponding Cuda versions to Cuda13

### Checklists

- [x] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.
- [x] verified LICENSE file has not changed
- [x] updated jucx in poms
- [x] updated ucx in blossom dockerfile
- [x] opened MR for doc updates
- [x] updated Dockerfile examples
- [x] tested with Cuda13, arm build, `ucx_perftest` with `-m cuda`

---------

Signed-off-by: Zach Puller <zpuller@nvidia.com>