Remove nested quotes and double quotes [skip ci] (#13580)
Fixes #13554 

### Description

The test script for iceberg tests have some bugs in it as it has
unnecessary quotes and double quotes in other quotes or double quotes.
They are syntactically incorrect and causing test failures. Two patterns
are fixed.

- For the config prefixed with `env`, the config value should not be
double quoted as the whole config key-value pair is quoted.
- The default value should not be quoted or double quoted.

I tested this fix manually on my workstation. The classNotFound error is
gone with the fix, but the tests did not pass successfully. The error
seems because of some missing settings for the rest catalog server, but
I'm not sure what I'm missing unfortunately.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>