Use `RapidsFileIO` for writing data. (#13501)
Fixes #13471 .

### Description

This is blocked by https://github.com/NVIDIA/spark-rapids-jni/pull/3768.
In this pr we implemented output related interface for `HadoopFileIO`
and `IcebergFileIO`, and use them in `ColumnarOutputWriter`.

### Checklists


- [x] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: liurenjie1024 <liurenjie2008@gmail.com>