Migrate CollectTimeIterator to use new NvtxIdWithMetrics class (#12900)
We are in the process of migrating all of our nvtx ranges to use the new
documented Nvtx api. Many of the existing ranges are already wrapped in
other classes such as the `NvtxWithMetrics` class. This pr establishes a
replacement for `NvtxWithMetrics` called `NvtxIdWithMetrics` which
provides the same wrapping functionality, but using the documented Nvtx
class as input. Then, the PR begins migrating some of the many uses of
this wrapper, namely the ranges that go through the
`CollectTimeIterator`

---------

Signed-off-by: Zach Puller <zpuller@nvidia.com>