Migrate most remaining NvtxRanges to use NvtxId/NvtxRangeWithDoc (#13614)
### Description

This PR migrates most the remaining NvtxRanges to use the documented
NvtxId wrappers, which will get auto generated into a docs page.

### Checklists

- [x] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Zach Puller <zpuller@nvidia.com>