Use assertInTests for cases with potential side effects, or expensive calls [databricks] (#13584)
Contributes to https://github.com/NVIDIA/spark-rapids/issues/13310

### Description

This PR adds a function `assertInTests` that detects whether the JVM is
under a test (integration tests, scala tests). We then use this function
in a few key areas were `assert` was being used directly when calling
`.hasNext` (could have side effects) and a few other places that could
be expensive.

If all goes well here, we could replace all other `assert` calls with
`assertInTests`, but I consider this PR enough to mostly deal with the
issue that was raised in #13310. It would just be cleaner if we always
used the same function, rather than have to think about different
asserts for different cases.

The code I generated mostly with the help of AI. I did not test the
bloop case, I could use help verifying that running the unit tests via
bloop would be covered with the checks performed in `AssertionUtils`.

I'll run NDS with this also, but I wanted to kick CI.

### Checklists

- [x] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Alessandro Bellina <abellina@nvidia.com>