Move AssertUtils to the api submodule (#13654)
Fixes https://github.com/NVIDIA/spark-rapids/issues/13647

### Description

Ok, so I believe the integration tests locally and in pre-merge are
running against a specific non-parallel-world jar, so all classes are in
the expected namespaces (not scribbled away over at
sparkver/com/nvidia..), but the nightly doesn't seem to be doing that.

I had added a class that I need to be able to read from the
`integration_tests` submodule. And that seems to be getting shimmed.

Here are my changes to help fix:

- I moved it to the "api" submodule, since that's not shimmed.
- I also added a dependency to `integration_test` to get the api
submodule.

I am having trouble testing locally (network). I'd like @gerashegalov's
opinion on this patch, as he knows the build system better than anyone.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Alessandro Bellina <abellina@nvidia.com>