Add allow_non_gpu_delta_write_if to conditionally allow CPU fallback for delta writes (#13643)
Fixes https://github.com/NVIDIA/spark-rapids/issues/13641.

### Description

The tests in question are failing because the dynamic partition
overwrite is supported only for Delta 3.3
(https://github.com/NVIDIA/spark-rapids/pull/13573). To allow the
fallback of Delta writes in older versions, I added a new
`allow_non_gpu_delta_write_if` marker. This new marker is different from
`allow_non_gpu` and can be used only for Delta writes. Note that the
Delta writes do not appear in the query plan and `allow_non_gpu` is not
usable in that case. The fix passed all delta tests with Spark 3.3.0 and
Delta 2.3.0.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>