Add missing javadocs in delta-33x-40x refactored code [skip ci] (#13692)
Contributes to https://github.com/NVIDIA/spark-rapids/issues/13339

### Description
 
Missed copying the javadocs from delta-33x files when we refactored it
to delta-33x-40x directory in these PR's
https://github.com/NVIDIA/spark-rapids/pull/13642
https://github.com/NVIDIA/spark-rapids/pull/13644

Compared and added the missing javadocs. 

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

Signed-off-by: Niranjan Artal <nartal@nvidia.com>