Add StatsExprShim in databricks shim [databricks] (#13701)
Fixes https://github.com/NVIDIA/spark-rapids/issues/13694

### Description

While working on https://github.com/NVIDIA/spark-rapids/pull/13666, we
added StatsExprShim to all non-DB shims but missed adding it to DB shim.
This PR adds it to DB shim
This should fix compile errors of DB builds

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

Signed-off-by: Niranjan Artal <nartal@nvidia.com>