Use new API to do Iceberg partition. (#13688)
Fixes #13679.

Use new API to do Iceberg partition.

Signed-off-by: Chong Gao <res_life@163.com>

Depends on 
* https://github.com/rapidsai/cudf/pull/20391

### Checklists

- [x] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
- [x] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.
Perf test is not needed, because it's obvious that new API is fast. New
API reduces multiple kernels into one kernel.

---------

Signed-off-by: Chong Gao <res_life@163.com>
Co-authored-by: Chong Gao <res_life@163.com>