Ensure that task threads are the only ones registered as dedicated (#13775)
Fixes https://github.com/NVIDIA/spark-rapids/issues/13774

### Description

This fix makes sure we don't register pool threads as "dedicated" or
task threads.

Note that we can probably clean this class a bunch, but I am putting
this up as a minimal change for now.

### Checklists
- [x] This PR has added documentation for new or modified features or
behaviors.
- [ ] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

Signed-off-by: Alessandro Bellina <abellina@nvidia.com>