Add Spark 3.5.7 support [databricks] (#13742)
Fixes #13470.

### Description

This PR adds Spark 3.5.7 support to the plugin. As the latest version in
the 3.5 line, we should support it. This change should be tested by
existing tests.

### Checklists

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>
Co-authored-by: Dev Bot <devbot@local>