Add missing handling for the 0 input partition count in computing partition count in GpuOptimizeWriteExchangeExec (#13780)
Fixes #13779.

`GpuOptimizeWriteExchangeExec` computes the `actualNumPartitions` based
on the partition count of its input. The computed `actualNumPartitions`
is used to dynamically optimize the partitioning. This logic is
currently missing handling of the case when the input partition count is
0, which can cause the ArithmeticException error. This PR adds a proper
handling of the case.

- [ ] This PR has added documentation for new or modified features or
behaviors.
- [x] This PR has added new tests or modified existing tests to cover
new code paths.
(Please explain in the PR description how the new code paths are tested,
such as names of the new/existing tests that cover them.)
- [ ] Performance testing has been performed and its results are added
in the PR description. Or, an issue has been filed with a link in the PR
description.

---------

Signed-off-by: Jihoon Son <ghoonson@gmail.com>